Add files via upload
Input Format

Input begins with three integers, A, B, and M, subject to the following constraints:

1 <= A, B <= 100,000
1 <= M <= 2,147,000,000

Next is a line with A positive integers, each less than or equal to 2,147,000,000. These integers are the prices of items in the first category.

Next is a line with B positive integers, each less than or equal to 2,147,000,000. These integers are the prices of items in the second category.

Sample Input

3 2 50

25 10 49

Sample Output

51
41 27